# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Topping CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CLIENTS ────────────────────────────────────────────────────────────────
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── LEADS ──────────────────────────────────────────────────────────────────
...
  /compensation/calculate:
    post:
      operationId: calculateCompensation
      tags: [compensation]
      summary: Deterministic commission / bonus calculation for a customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompensationInput"
      responses:
        "200":
          description: Full payout breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationResult"
  /compensation/ask:
    post:
      operationId: askCompensation
      tags: [compensation]
      summary: Free-form compensation question answered by the Compensation AI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompensationAskInput"
      responses:
        "200":
          description: AI answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationAskResponse"
...
components:
  schemas:
...
    Company:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        industry:
          type: string
        website:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        website:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string

    # ─── CLIENT ───────────────────────────────────────────────────────────────
    Client:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, prospect]
        monthlyRevenue:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        companyId:
          type: integer
        status:
          type: string
        monthlyRevenue:
          type: number
        notes:
          type: string
        ownerId:
          type: integer
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        companyId:
          type: integer
        status:
          type: string
        monthlyRevenue:
          type: number
        notes:
          type: string
        ownerId:
          type: integer

    # ─── LEAD ─────────────────────────────────────────────────────────────────
...
    CompensationGates:
      type: object
      required:
        [firstOrder, crmVerified, invoicePaid, noDispute, nodeCompliance, leadSourceLogged]
      properties:
        firstOrder:
          type: boolean
        crmVerified:
          type: boolean
        invoicePaid:
          type: boolean
        noDispute:
          type: boolean
        nodeCompliance:
          type: boolean
        leadSourceLogged:
          type: boolean
    CompensationInput:
      type: object
      required: [monthlyRevenue, scenario, gates]
      properties:
        customerName:
          type: string
        monthlyRevenue:
          type: number
          description: First month's customer revenue (USD)
        scenario:
          type: string
          enum: [A, B, C, D]
        leadGenerator:
          type: string
          enum: [Arshia, Shayan, none]
        fieldSalesPerson:
          type: string
        closerPerson:
          type: string
        newCustomersThisMonth:
          type: integer
          description: Customers this lead-gen activated this month (monthly performance bonus)
        strategicType:
          type: string
          enum: [none, over_10k, multi_branch, long_term]
        founderApproved:
          type: boolean
        gates:
          $ref: "#/components/schemas/CompensationGates"
    CompensationPayoutLine:
      type: object
      required: [role, person, amount, status, basis]
      properties:
        role:
          type: string
        person:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [confirmed, pending, blocked]
        basis:
          type: string
    CompensationGateResult:
      type: object
      required: [gate, passed]
      properties:
        gate:
          type: string
        passed:
          type: boolean
    CompensationResult:
      type: object
      required:
        [
          customerName,
          monthlyRevenue,
          sizeLabel,
          commissionRate,
          commissionPool,
          scenario,
          payouts,
          totalAcquisitionCost,
          pendingTotal,
          operationalProfit,
          acquisitionPctOfProfit,
          companyNetProfit,
          blocked,
          gateResults,
          flags,
        ]
      properties:
        customerName:
          type: string
        monthlyRevenue:
          type: number
        sizeLabel:
          type: string
        commissionRate:
          type: number
        commissionPool:
          type: number
        scenario:
          type: string
          enum: [A, B, C, D]
        payouts:
          type: array
          items:
            $ref: "#/components/schemas/CompensationPayoutLine"
        totalAcquisitionCost:
          type: number
        pendingTotal:
          type: number
        operationalProfit:
          type: number
        acquisitionPctOfProfit:
          type: number
        companyNetProfit:
          type: number
        blocked:
          type: boolean
        gateResults:
          type: array
          items:
            $ref: "#/components/schemas/CompensationGateResult"
        flags:
          type: array
          items:
            type: string
    CompensationAskInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
          description: Free-form compensation question (Farsi/Finglish or English)
    CompensationAskResponse:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
          description: The Compensation AI's natural-language answer